// backend/src/config/env.d.ts
export declare const env: {
    NODE_ENV: "development" | "production" | "test";
    PORT: string;
    JWT_SECRET: string;
    STORAGE_MODE: "in-memory" | "sql";
    DATABASE_DIALECT: "mysql" | "postgres";
    DATABASE_URL?: string | undefined;
    FRONTEND_URL: string;
};
//# sourceMappingURL=env.d.ts.map